pyproject.toml
[tool.poetry]
name = "lfk.im-git"
version = "0.1.0"
description = ""
authors = ["Jeff Triplett <jeff.triplett@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-frontmatter = "^0.5.0"
python-slugify = "^4.0.0"
click = "^7.1.1"
sheetfu = "^1.5.2"
typesystem = "^0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
